Splitting data_preparation.py file into menu and functions
